Use symmetrical hash style
If a hash's opening brace is on the same line as the first element of
the hash, then the closing brace should be on the same line as the last
element of the hash.

If a hash's opening brace is on the line above the first element of the
hash, then the closing brace should be on the line below the last
element of the hash.